Update flowframe.py
I make it so the flow frame can be used as an ordinary frame, it only starts working when addWidget is called, organizeWithPlace, or organizeWithPack is called.  This way you can substitute this into code using ordinary frame methods, and then when added functionality is needed, you use one of those 3 methods.